pyproject.toml
[tool.poetry]
name = "git_cd"
version = "0.1.0"
description = ""
authors = ["Danyiel Colin <danyiel5978@gmail.com>"]
readme = "README.md"
packages = [{include = "gitcd"}]

[tool.poetry.dependencies]
python = "^3.11"
gitpython = "^3.1.31"

[tool.poetry.group.dev.dependencies]
pynvim = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
git-cd = "gitcd.main:main"
